requirements.txt
python3 - 64 bit and latest version!
pytorch
adversarial robustness toolbox (ART)
NumPy
pyqt5-tools
maybe tensorflow/keras
DB Browser